Improve description and remove promises
