+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "tfg-event-context",
-  "version": "0.5.7",
+  "version": "0.6.0",
   "description": "Universal Event context for JavaScript",
   "main": "index.js",
   "jsnext:main": "index.es.js",
