+++ src/index.js
@@ -1,6 +1,6 @@
 export * from './context';
 import { patch as patchNatives } from './natives';
-import { patch as patchPromise } from './promise';
+import { patch as patchPromise } from './async';
 
 patchNatives();
 if (typeof Promise !== 'undefined') {
